+++ src/core/swap/swap-api.ts
@@ -76,6 +76,22 @@ export async function fetchSwapQuotes(
           error => {
             pendingIds.delete(pluginId)
             log.warn(`${pluginId} gave swap error: ${String(error)}`)
+            // Log unknown errors:
+            if (isUnknownSwapError(error)) {
+              log.crash(`Unknown swap error: ${String(error)}`, {
+                error,
+                swapPluginId: pluginId,
+                request: {
+                  // Stringify to include "null"
+                  fromToken: String(request.fromTokenId),
+                  fromWalletType: request.fromWallet.type,
+                  // Stringify to include "null"
+                  toToken: String(request.toTokenId),
+                  toWalletType: request.toWallet.type,
+                  quoteFor: request.quoteFor
+                }
+              })
+            }
             throw error
           }
         )
@@ -231,4 +247,10 @@ function rankError(error: unknown): number {
   if (asMaybeSwapPermissionError(error) != null) return 3
   if (asMaybeSwapCurrencyError(error) != null) return 2
   return 1
+}
+
+function isUnknownSwapError(error: unknown): boolean {
+  const isKnownError = rankError(error) > 1
+  // NOTE: Add more error filtering here as we decide to filter out noise.
+  return !isKnownError
 }